Create golden_runner package
