+++ golden_runner/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: "f8afcd5aa01b3ae6c55cb6e4c9fa4171e27a92f6"
+  channel: "master"
+
+project_type: package
