+++ golden_runner/lib/golden_runner.dart
@@ -1,0 +1,3 @@
+library golden_runner;
+
+export 'src/commands.dart';
