+++ super_editor/pubspec.yaml
@@ -38,10 +38,10 @@ dependencies:
   flutter_test_robots: ^0.0.17
 
 dependency_overrides:
-#  # Override to local mono-repo path so devs can test this repo
-#  # against changes that they're making to other mono-repo packages
-#  attributed_text:
-#    path: ../attributed_text
+  #  # Override to local mono-repo path so devs can test this repo
+  #  # against changes that they're making to other mono-repo packages
+  #  attributed_text:
+  #    path: ../attributed_text
   super_text_layout:
     path: ../super_text_layout
 #
@@ -59,6 +59,8 @@ dev_dependencies:
   meta: ^1.8.0
   args: ^2.3.1
   path: ^1.8.3
+  golden_runner:
+    path: ../golden_runner
 
 flutter:
   # no Flutter configuration
